backend/requirements.txt
fastapi
uvicorn[standard]
websockets
openai
anthropic
httpx
python-dotenv
pydantic
aiosqlite
SpeechRecognition
pydub
transformers
torch
sentencepiece
openpyxl
python-multipart
pyjwt
bcrypt
reportlab
langgraph
google-generativeai
langchain-core
imageio-ffmpeg
